GetFees function, readme updates, satoshi Math.floor update
