+++ src/helper/index.js
@@ -1,7 +1,9 @@
 const signTransaction = require('./signTransaction')
+const getFeeAndInput = require('./calculateFeeAndInput')
 const utils = require('./utils/index')
 
 module.exports = {
     signTransaction,
-    utils
+    utils,
+    getFeeAndInput
 }
