+++ src/helper/signTransaction.js
@@ -2,35 +2,16 @@ const axios = require('axios')
 const bitcore = require("bitcore-lib")
 const { SATOSHI } = require("../config/index")
 
+const getFeeAndInput = require('./calculateFeeAndInput')
 
 async function signTransaction(from, to, amountToSend, URL, privateKey) {
 
     const satoshiToSend = amountToSend * SATOSHI;
-    let fee = 0;
-    let inputCount = 0;
-    let outputCount = 2;
-    const utxos = await axios.get(URL);
 
     const transaction = new bitcore.Transaction();
-    let totalAmountAvailable = 0;
 
-    let inputs = [];
-    utxos.data.data.txs.forEach(async (element) => {
-        let utxo = {};
-        utxo.satoshis = Math.floor(Number(element.value) * SATOSHI);
-        utxo.script = element.script_hex;
-        utxo.address = utxos.data.data.address;
-        utxo.txId = element.txid;
-        utxo.outputIndex = element.output_no;
-        totalAmountAvailable += utxo.satoshis;
-        inputCount += 1;
-        inputs.push(utxo);
-    });
+    const {totalAmountAvailable, inputs, fee} = await getFeeAndInput(URL)
 
-    transactionSize = inputCount * 146 + outputCount * 34 + 10 - inputCount;
-    // Check if we have enough funds to cover the transaction and the fees assuming we want to pay 20 satoshis per byte
-
-    fee = transactionSize * 20
     if (totalAmountAvailable - satoshiToSend - fee < 0) {
         throw new Error("Balance is too low for this transaction");
     }
@@ -39,7 +20,7 @@ async function signTransaction(from, to, amountToSend, URL, privateKey) {
     transaction.from(inputs);
 
     // set the recieving address and the amount to send
-    transaction.to(to, satoshiToSend);
+    transaction.to(to, Math.floor(satoshiToSend));
 
     // Set change address - Address to receive the left over funds after transfer
     transaction.change(from);
